+++ eyes.appium.java/src/test/java/com/applitools/eyes/appium/ios/IOSTestSetup.java
@@ -13,7 +13,7 @@ public abstract class IOSTestSetup extends TestSetup {
     protected void setCapabilities() {
         super.setCapabilities();
         capabilities.setCapability("platformName", "iOS");
-        capabilities.setCapability("deviceName", "iPhone 8");
+        capabilities.setCapability("device", "iPhone 8");
         capabilities.setCapability("os_version", "12");
         capabilities.setCapability("automationName", "XCUITest");
         capabilities.setCapability("newCommandTimeout", 300);